+++ NEWS.md
@@ -4,6 +4,7 @@ PENDING: dataiter 0.27
 * `DataFrame`: Fix error message when column not found
 * `DataFrame.aggregate`: Speed up
 * `DataFrame.read_csv`: Try to avoid mixed types
+* `ListOfDicts.write_csv`: Use minimal quoting
 * `Vector.rank`: Rewrite, add `method` argument
 * `*.read_*`: Rename fname argument path
 * `*.write_*`: Rename fname argument path